MDEV-24302 RESET MASTER hangs as Innodb does not report on binlog checkpoint
The current commit addresses a specific flavor of the bug which
is not about races of the preceding commit.

Binlog checkpoint notification consumers such as RESET MASTER
receive a notification and proceed when lsn_0 at the time of the
request is finally gets flushed
  flush_lsn >= lsn_0
The bug situation was that when lsn_0 reflects a write by an internal innodb trx
and RESET MASTER was not followed by any more user transaction
it would hang.